Update issuer verify tests.
